+++ jsonpath/env.py
@@ -252,6 +252,10 @@ class JSONPathEnvironment:
         self.function_extensions["match"] = function_extensions.Match()
         self.function_extensions["search"] = function_extensions.Search()
         self.function_extensions["value"] = function_extensions.value
+        self.function_extensions["isinstance"] = function_extensions.IsInstance()
+        self.function_extensions["is"] = self.function_extensions["isinstance"]
+        self.function_extensions["typeof"] = function_extensions.TypeOf()
+        self.function_extensions["type"] = self.function_extensions["typeof"]
 
     def validate_function_extension_signature(
         self, token: Token, args: List[Any]