+++ jsonpath/function_extensions/__init__.py
@@ -1,18 +1,22 @@
 # noqa: D104
 from .arguments import validate
 from .count import Count
+from .is_instance import IsInstance
 from .keys import keys
 from .length import length
 from .match import Match
 from .search import Search
+from .typeof import TypeOf
 from .value import value
 
 __all__ = (
     "Count",
-    "Match",
-    "Search",
-    "value",
+    "IsInstance",
     "keys",
     "length",
+    "Match",
+    "Search",
+    "TypeOf",
     "validate",
+    "value",
 )